pyproject.toml
[project]
name = "telegram-webhook"
version = "0.1.0"
description = "Telegram bot webhook → local PC copilot bridge"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-dotenv>=1.0",
    "httpx>=0.28",
]
